Add `DB::ColumnsScatter::scatter` with bytes-budget input batching
Implements the scatter half of the radix-shuffle API proposed in issue #105936
together with its consumer in `BufferedShardByHashTransform`.

`DB::ColumnsScatter::scatter(source_columns, pids_per_source, per_shard_rows, num_shards)`
is a batched, zero-virtual-dispatch physical-split primitive. The caller supplies
one `IColumn` from each of B pending chunks (same column-position across a batch),
along with a per-source span of `UInt32` partition-ids and the precomputed
per-shard row counts. An `O(1)` dispatch table keyed on the column's type index
selects the per-type kernel; there is no virtual call and no linear type switch on
the hot path.

Each typed kernel (`ColumnVector`, `ColumnDecimal`, `ColumnFixedString`,
`ColumnString`, `ColumnNullable`, `ColumnTuple`) follows the same shape: a single
`O(P)` setup loop that allocates destinations via `reserve_exact`, initialises a
stack-resident write-pointer cache and commits the final size via
`resize_assume_reserved`, then a hot inner loop `*wp[pids[j]]++ = src[j]` across
all B sources. `ColumnString` keeps the chars, cursor and offsets scatter
single-pass and uses `memcpySmallAllowReadWriteOverflow15` (safe because both
source and destination `PaddedPODArray<UInt8>` carry right-padding).
`ColumnNullable` scatters the null map via the `UInt8` kernel then recurses into
the nested column; `ColumnTuple` recurses per element. Types without a fast path
(`ColumnArray`, Map, LowCardinality, Sparse, ...) and `ColumnConst` fall back to
the legacy `IColumn::scatter` virtual.

`DB::ColumnsScatter::computeHistogram` returns the per-shard row counts as
`std::span<const UInt32>`; `flushBatch` computes them once per flush and passes
them to every `scatter` call, letting each kernel skip its internal pids re-scan.
The scatter kernels and `per_shard_rows` consistently use `UInt32` for per-shard
counts, matching the `UInt32` partition-id width.

Consumer - `BufferedShardByHashTransform` input batching:

The new setting `shard_by_hash_input_batch_bytes` (default 0 = legacy per-chunk
path) enables input batching. When non-zero the transform stashes input chunks in
`pending_input`, accumulating each chunk's `UInt32` pids into a shared buffer.
Once the accumulated input bytes reach the budget, or on input exhaustion,
`flushBatch` calls `ColumnsScatter::scatter` once per column-position across all
pending chunks, producing exact-sized per-shard `MutableColumns` with no per-chunk
KxP `IColumn` allocation.

The flush is bounded by accumulated bytes rather than a row count. Counting bytes
keeps each flush short and its per-shard destination buffers small for wide rows
(a long single-threaded flush starves the parallel downstream
`AggregatingTransform` instances and faults large transient buffers), while narrow
rows still accumulate until the budget is reached. Chunks are never split, so a
batch may overshoot the budget by at most one input chunk's bytes.

The `selector` member is a `std::variant<IColumn::Selector, PaddedPODArray<UInt32>>`:
the `UInt64` `IColumn::Selector` alternative serves the legacy path (`mapToRange`
writes `UInt64` directly, no widening copy), the `PaddedPODArray<UInt32>`
alternative is the shared batched pids buffer. `mapToRange` gains dual `UInt32` /
`UInt64` write overloads sharing the `MULTITARGET_FUNCTION_X86_V4_V3` machinery.

Measured performance:

Sweep on an AWS `c7i.metal-24xl` (Intel Xeon Platinum 8488C, Sapphire Rapids,
1 socket, 48 cores / 96 threads, ~188 GiB, `ap-south-2`). Workload: 100M-row
`SELECT keys, count() ... GROUP BY keys` over `Memory` tables with
`enable_sharding_aggregator=1` and `max_threads=16`. Key/row width was swept
8-64 B (1-8 `UInt64` key columns, spanning the `key64`/`key128`/`key256`/
serialized aggregation methods) and cardinality over 100 / 1e5 / 1e7 distinct
keys. Wall time is the median of 9-15 interleaved iterations vs the legacy
per-chunk path (`shard_by_hash_input_batch_bytes = 0`). Ranges below span the
three cardinalities at the recommended 2 MiB budget:

  -  8 B keys: -9% to -25%   (largest gain: low cardinality, -25%)
  - 16 B keys: neutral to -10%  (high cardinality is ~flat at 2 MiB; smaller
               budgets reach -15% there)
  - 32 B keys: -8% to -12%
  - 64 B keys: -7% to -10% at medium/high cardinality, neutral at low
               cardinality (within run-to-run noise)

The byte budget must stay bounded: an oversized 8 MiB budget reintroduces the
wide-row regression it exists to prevent (+16% on low-cardinality 64 B keys, and
4 MiB already starts to erode the wide-row wins), because each flush again grows
long enough to starve the downstream aggregators. 2 MiB is the robust optimum
across the sweep (best worst-case and best narrow-row peak), hence the documented
recommended starting value.

Also:
- `BuildQueryPipelineSettings` exposes `shard_by_hash_input_batch_bytes`.
- `benchmark_columns_scatter.cpp`: RFC sweep over K and P.
- `gtest_columns_scatter.cpp`: equivalence vs legacy for all fast-path types
  (including `Time64`), `ColumnArray` / `ColumnConst` fallback coverage, and
  per-key count assertions that catch misrouted rows.
- `04278_04101_sharded_aggregation_batched_input.sql`: batched vs legacy
  per-key-identical `GROUP BY` results across key types.

Closes #105936

+++ src/Common/MapToRange.h
@@ -20,6 +20,9 @@ namespace DB
 /// vpmuludq (1 µop, 0.5c throughput) instead of vpmullq (3 µops).
 ///
 /// range_size must fit in uint32_t (any realistic partition count does).
-void mapToRange(const UInt32 * hashes, size_t n, UInt32 range_size, UInt64 * result);
+///
+/// UInt32 output — for use with ColumnsScatter::scatter.
+/// Halves selector bandwidth in the hot scatter inner loop.
+void mapToRange(const UInt32 * hashes, size_t n, UInt32 range_size, UInt32 * result);
 
 } // namespace DB
